qrant_rag_w_memory/requirements.txt
qdrant-client
docling
fastembed
google-generativeai
python-dotenv